Create create-graph.py
The following code calls createGraph(numVertices, numGraphs, alaramOn), which generates a total number of numGraphs graphs, starting with a linear graph with n = numVertices vertices, saving it in a csv file, then in every step it read the previous csv file, generates a matrix of it, densifies it by adding numNewEdgesPerGraph edges and restores the denser graph in a new csv file.
File names have the following form: numV ertices.graphNum.csv. The higher graphNum is the denser the graph is.